+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 
 # CHANGELOG
 
+## version 1.2
+
+- 检测抛出配置文件中路径配置项结尾缺失`/`的提示语
+- 更改配置文件传入的方式，现在传入配置文件路径需要在前面增加`-c`或者`--config`
+
+    ```shell
+    UpdateBackupHelper.exe -c=./config.json # 第一种方式
+    UpdateBackupHelper.exe -c ./config.json # 第二种方式
+    ```
+
+- 另外，新增两个传入参数，优先级高于配置文件中的配置项，分别是`last_commit_num`、`author`配置项
+
+    - `last_commit_num` 最后提交的git次数，`-q`或者`--quantity`
+    - `author` 提交的git作者，`-a`或者`--author`
+
+    ```shell
+    UpdateBackupHelper.exe -c ./config.json -a GHBJayce -q 6 # 第二种方式
+    ```
+
 ## version 1.1
 
 - [x] 拷贝不存在（新增/删除）的文件时，使用捕获异常并给出提示的方式告知用户